refactor: extract emoji_picker / contacts / profile / forward overlays (#357)
Tenth slice. Four medium picker-style overlays bundled:

- draw_emoji_picker  (~120 lines) - category tabs + scrolling emoji grid
- draw_contacts      (~92 lines)  - type-to-filter contacts list
- draw_profile       (~90 lines)  - 4-field profile editor with Save button
- draw_forward       (~67 lines)  - filterable conversation picker

CATEGORIES (from crate::domain) and crate::list_overlay are now only
used by overlay submodules, so both imports moved out of ui/mod.rs.

ui/mod.rs: 3938 -> 3564 lines.
