+++ src/ui/overlays/mod.rs
@@ -7,12 +7,16 @@
 
 pub(super) mod about;
 pub(super) mod action_menu;
+pub(super) mod contacts;
 pub(super) mod customize;
 pub(super) mod delete_confirm;
+pub(super) mod emoji_picker;
+pub(super) mod forward;
 pub(super) mod keybindings;
 pub(super) mod message_request;
 pub(super) mod pin_duration;
 pub(super) mod poll_vote;
+pub(super) mod profile;
 pub(super) mod reaction_picker;
 pub(super) mod settings;
 pub(super) mod settings_profile;
